# ProductTraceability_2026/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Agricultural Traceability System API
  version: 1.0.0
  description: API specification for the agricultural product traceability system.
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: http://localhost:8888
paths:
...
  /api/enterprise/audit/list:
    get:
      summary: "[Full] Enterprise audit list"
      tags: [Enterprise]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: pageNum, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 10 } }
        - { name: auditStatus, in: query, schema: { type: integer }, description: "0-pending, 1-approved, 2-rejected" }
        - { name: enterpriseType, in: query, schema: { type: integer }, description: "1-farm, 2-processor, 3-inspector" }
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated enterprise list
...
  /api/enterprise/audit/approve:
    post:
      summary: "[Full] Approve or reject enterprise"
      tags: [Enterprise]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enterpriseId, auditStatus]
              properties:
                enterpriseId: { type: integer }
                auditStatus: { type: integer, description: "1-approve, 2-reject" }
                auditRemark: { type: string }
      responses:
        '200':
          description: Success
...
  /api/image-audit/submit:
    post:
      summary: "✅ Submit image change for audit"
      tags: [Image-Audit]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enterpriseId, fieldName, newValue]
              properties:
                enterpriseId: { type: integer }
                userId: { type: integer }
                fieldName: { type: string, description: "logo, business_license, production_license" }
                oldValue: { type: string }
                newValue: { type: string }
      responses:
        '200':
          description: Audit record created
  /api/image-audit/list:
    get:
      summary: "✅ Image audit list (admin)"
      tags: [Image-Audit]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: pageNum, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 10 } }
        - { name: auditStatus, in: query, schema: { type: integer }, description: "0-pending, 1-approved, 2-rejected" }
        - { name: enterpriseId, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Paginated audit list
  /api/image-audit/approve:
    post:
      summary: "✅ Approve or reject image audit"
      tags: [Image-Audit]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, auditStatus]
              properties:
                id: { type: integer }
                auditStatus: { type: integer, description: "1-approve, 2-reject" }
                remark: { type: string }
      responses:
        '200':
          description: Success
  /api/image-audit/pending-count:
    get:
      summary: "✅ Get pending image audit count for enterprise"
      tags: [Image-Audit]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: enterpriseId, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Pending audit count
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  data: { type: integer, description: "Number of pending audits" }
...
  /api/edu/article/published:
    get:
      summary: "✅ Get published articles (public)"
      tags: [Education]
      parameters:
        - name: enterpriseId
          in: query
          schema: { type: integer }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of published articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/EduArticle' }
  /api/edu/article/detail/{id}:
    get:
      summary: "✅ Get article detail (public)"
      tags: [Education]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Article detail
  /api/edu/article/view/{id}:
    put:
      summary: "✅ Increment article view count (public)"
      tags: [Education]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Success
  /api/edu/article/list:
    get:
      summary: "✅ Article list with pagination (admin)"
      tags: [Education]
      parameters:
        - name: current
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: enterpriseId
          in: query
          schema: { type: integer }
        - name: category
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: integer }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated article list
  /api/edu/article/add:
    post:
      summary: "✅ Add a new article (admin)"
      tags: [Education]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EduArticle' }
      responses:
        '200':
          description: Success
  /api/edu/article/update:
    put:
      summary: "✅ Update an article (admin)"
      tags: [Education]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EduArticle' }
      responses:
        '200':
          description: Success
  /api/edu/article/delete/{id}:
    delete:
      summary: "✅ Delete an article (admin)"
      tags: [Education]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Success
  /api/edu/article/toggle-status/{id}:
    put:
      summary: "✅ Toggle article publish/draft (admin)"
      tags: [Education]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Success

  # ==========================================
  # Enterprise Registration (Public)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EduArticle:
      type: object
      properties:
        id: { type: integer }
        enterpriseId: { type: integer, description: "Enterprise ID, null for platform-level" }
        category: { type: string, description: "farming/vaccine/process/hygiene/quarantine/blockchain/safety" }
        title: { type: string }
        author: { type: string }
        summary: { type: string }
        coverUrl: { type: string }
        content: { type: string, description: "JSON array of content blocks" }
        viewCount: { type: integer }
        sortOrder: { type: integer }
        status: { type: integer, description: "0-draft, 1-published" }
        publishTime: { type: string, format: date-time }